% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named-palettes.R
\name{named_palettes}
\alias{named_palettes}
\alias{pal_meadow}
\alias{pal_atelier}
\alias{pal_clinical}
\alias{pal_spectrum}
\alias{pal_pastel}
\alias{pal_earth}
\alias{pal_midnight}
\alias{pal_floral}
\alias{pal_coastal}
\alias{pal_harvest}
\alias{pal_aurora}
\alias{pal_ocean}
\alias{pal_ember}
\alias{pal_forest}
\alias{pal_orchid}
\alias{pal_slate}
\alias{pal_sunrise}
\alias{pal_glacier}
\alias{pal_berry}
\alias{pal_sand}
\alias{pal_balance}
\alias{pal_temperature}
\alias{pal_earthsky}
\alias{pal_rosepine}
\alias{pal_lagoon}
\alias{pal_violetgold}
\title{Named ggpalettes palette functions}
\usage{
pal_meadow(alpha = 1, reverse = FALSE)

pal_atelier(alpha = 1, reverse = FALSE)

pal_clinical(alpha = 1, reverse = FALSE)

pal_spectrum(alpha = 1, reverse = FALSE)

pal_pastel(alpha = 1, reverse = FALSE)

pal_earth(alpha = 1, reverse = FALSE)

pal_midnight(alpha = 1, reverse = FALSE)

pal_floral(alpha = 1, reverse = FALSE)

pal_coastal(alpha = 1, reverse = FALSE)

pal_harvest(alpha = 1, reverse = FALSE)

pal_aurora(alpha = 1, reverse = FALSE)

pal_ocean(alpha = 1, reverse = FALSE)

pal_ember(alpha = 1, reverse = FALSE)

pal_forest(alpha = 1, reverse = FALSE)

pal_orchid(alpha = 1, reverse = FALSE)

pal_slate(alpha = 1, reverse = FALSE)

pal_sunrise(alpha = 1, reverse = FALSE)

pal_glacier(alpha = 1, reverse = FALSE)

pal_berry(alpha = 1, reverse = FALSE)

pal_sand(alpha = 1, reverse = FALSE)

pal_balance(alpha = 1, reverse = FALSE)

pal_temperature(alpha = 1, reverse = FALSE)

pal_earthsky(alpha = 1, reverse = FALSE)

pal_rosepine(alpha = 1, reverse = FALSE)

pal_lagoon(alpha = 1, reverse = FALSE)

pal_violetgold(alpha = 1, reverse = FALSE)
}
\arguments{
\item{alpha}{Opacity between 0 and 1.}

\item{reverse}{Reverse the palette.}
}
\value{
A function accepting the requested number of colours.
}
\description{
These convenience functions follow the \verb{pal_*()} interface popularized by
scientific palette packages.
}
