% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{gg_palette_names}
\alias{gg_palette_names}
\title{List available ggpalettes palettes}
\usage{
gg_palette_names(type = c("all", .palette_types))
}
\arguments{
\item{type}{Palette type: all, categorical, sequential, or diverging.}
}
\value{
A character vector containing the available palette names.
}
\description{
List available ggpalettes palettes
}
\examples{
gg_palette_names()
gg_palette_names("diverging")
}
