% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_colour_ggpalette}
\alias{scale_colour_ggpalette}
\alias{scale_color_ggpalette}
\alias{scale_fill_ggpalette}
\title{ggplot2 colour and fill scales using ggpalettes}
\usage{
scale_colour_ggpalette(name = "aurora", discrete = TRUE, direction = 1, ...)

scale_color_ggpalette(name = "aurora", discrete = TRUE, direction = 1, ...)

scale_fill_ggpalette(name = "aurora", discrete = TRUE, direction = 1, ...)
}
\arguments{
\item{name}{Palette name. See \code{\link[=gg_palette_names]{gg_palette_names()}}.}

\item{discrete}{Whether to construct a discrete scale. Set to \code{FALSE} for a
continuous gradient.}

\item{direction}{Palette direction: \code{1} or \code{-1}.}

\item{...}{Additional arguments passed to the underlying \code{ggplot2} scale.}
}
\value{
A \code{ggplot2} scale object.
}
\description{
ggplot2 colour and fill scales using ggpalettes
}
\examples{
library(ggplot2)
ggplot(iris, aes(Sepal.Length, Sepal.Width, colour = Species)) +
  geom_point(size = 2) +
  scale_colour_ggpalette("clinical")
}
